Fixed notification, by switching over to battery_info package (see description)
workmanager is running on a different isolate and because of that battery_plus can't be called, related to this issue: https://github.com/flutter/flutter/issues/13937
