+++ pubspec.yaml
@@ -37,10 +37,10 @@ dependencies:
   http: ^0.13.4
   geolocator: ^8.0.1
   skeleton_loader: ^2.0.0+4
-  battery_plus: ^2.0.2
   flutter_local_notifications: ^9.1.5
   shared_preferences: ^2.0.11
   workmanager: ^0.4.1
+  battery_info: ^1.0.8
 
 
   # The following adds the Cupertino Icons font to your application.
